finishedSolvedBug

+++ admin/blog/index.php
@@ -1,8 +1,14 @@
 <?php
 // Fetch blogs with their related categories and users
-$statement = $db->prepare("SELECT blogs.id, blogs.title, blogs.content, blogs.created_at, blogs.image, categories.name AS category_name, users.name AS user_name FROM blogs
-INNER JOIN categories ON blogs.category_id = categories.id 
-INNER JOIN users ON blogs.user_id = users.id");
+$statement = $db->prepare("
+    SELECT 
+        blogs.id, blogs.title, blogs.content, blogs.created_at, blogs.image, 
+        categories.name AS category_name, 
+        users.name AS user_name 
+    FROM blogs
+    INNER JOIN categories ON blogs.category_id = categories.id 
+    INNER JOIN users ON blogs.user_id = users.id
+");
 $statement->execute();
 $blogs = $statement->fetchAll(PDO::FETCH_OBJ);
 
@@ -23,10 +29,18 @@ if (isset($_POST['delete-btn'])) {
 
     // If deletion is successful, remove the image file
     if ($result) {
-        if ($blogState && $blogState->image) {
+        if ($blogState && $blogState->image && file_exists("../assets/blog-image/" . $blogState->image)) {
             unlink("../assets/blog-image/" . $blogState->image);
         }
-        echo "<script>sweetAlert('Deleted a blog', 'blogs')</script>";
+        // Redirect after success to avoid form resubmission issues
+        echo "<script>
+                Swal.fire('Deleted!', 'The blog has been deleted.', 'success').then(() => {
+                    window.location.href = 'index.php?page=blogs';
+                });
+              </script>";
+    } else {
+        // Handle deletion failure
+        echo "<script>Swal.fire('Error', 'Failed to delete the blog.', 'error');</script>";
     }
 }
 ?>
@@ -37,7 +51,9 @@ if (isset($_POST['delete-btn'])) {
         <div class="card shadow mb-4 col-md-12">
             <div class="card-header py-3 d-flex justify-content-between align-items-center">
                 <h6 class="m-0 font-weight-bold text-primary">Blogs List</h6>
-                <a href="index.php?page=blogs-create" class="btn btn-primary btn-sm"><i class="fa fa-plus" aria-hidden="true"></i> Add New</a>
+                <a href="index.php?page=blogs-create" class="btn btn-primary btn-sm">
+                    <i class="fa fa-plus" aria-hidden="true"></i> Add New
+                </a>
             </div>
             <div class="card-body">
                 <div class="table-responsive">
@@ -57,23 +73,31 @@ if (isset($_POST['delete-btn'])) {
                         <tbody>
                             <?php foreach ($blogs as $index => $blog): ?>
                                 <tr>
-                                    <td><?php echo $index + 1; ?></td>
+                                    <td><?= $index + 1; ?></td>
                                     <td>
-                                        <img src="../assets/blog-image/<?php echo $blog->image; ?>" class="img-fluid" alt="" style="width: 100px;">
+                                        <img src="../assets/blog-image/<?= htmlspecialchars($blog->image); ?>" class="img-fluid" alt="Blog Image" style="width: 100px;">
                                     </td>
-                                    <td><?php echo $blog->title; ?></td>
-                                    <td><?php echo $blog->category_name; ?></td>
+                                    <td><?= htmlspecialchars($blog->title); ?></td>
+                                    <td><?= htmlspecialchars($blog->category_name); ?></td>
                                     <td>
-                                        <div style="max-width:300px; max-height:200px; overflow:auto;"><?php echo $blog->content; ?></div>
+                                        <div style="max-width:300px; max-height:200px; overflow:auto;">
+                                            <?= htmlspecialchars($blog->content); ?>
+                                        </div>
                                     </td>
-                                    <td><?php echo $blog->user_name; ?></td>
-                                    <td><?php echo $blog->created_at; ?></td>
+                                    <td><?= htmlspecialchars($blog->user_name); ?></td>
+                                    <td><?= htmlspecialchars($blog->created_at); ?></td>
                                     <td>
                                         <form method="POST" class="d-flex">
-                                            <a href="index.php?page=blogs-edit&blog_id=<?php echo $blog->id; ?>" class="btn btn-primary m-1" title="Edit"><i class="far fa-edit"></i></a>
-                                            <input type="hidden" value="<?php echo $blog->id; ?>" name="blog_id">
-                                            <button class="btn btn-danger m-1" name="delete-btn" onclick="return confirm('Are you sure you want to delete this blog?')" title="Delete"><i class="fas fa-trash"></i></button>
-                                            <a href="index.php?page=blogs-comments&blog_id=<?php echo $blog->id; ?>" class="btn btn-primary m-1" title="Comment"><i class="fas fa-comment"></i></a>
+                                            <a href="index.php?page=blogs-edit&blog_id=<?= $blog->id; ?>" class="btn btn-primary m-1" title="Edit">
+                                                <i class="far fa-edit"></i>
+                                            </a>
+                                            <input type="hidden" value="<?= $blog->id; ?>" name="blog_id">
+                                            <button class="btn btn-danger m-1" name="delete-btn" onclick="return confirm('Are you sure you want to delete this blog?')" title="Delete">
+                                                <i class="fas fa-trash"></i>
+                                            </button>
+                                            <a href="index.php?page=blogs-comments&blog_id=<?= $blog->id; ?>" class="btn btn-primary m-1" title="Comments">
+                                                <i class="fas fa-comment"></i>
+                                            </a>
                                         </form>
                                     </td>
                                 </tr>
